fixing self-referencing

+++ run_test.sh
@@ -15,9 +15,11 @@ for d in */ ; do
 	fi
 	# running generic test
 	# refreshes and runs test.bats
-	cp ../../addTests.py .
-	./addTests.py
-	rm addTests.py
+	if [ ! -f "tests.bats" ]; then
+		cp ../../addTests.py .
+		./addTests.py
+		rm addTests.py
+	fi
 	bats tests.bats
 	restOfBats=$(./findRelevant.py -e .bats tests.bats) # attempts to find other .bats files
 	if [ $? -eq 0 ]; then
